-Added empty input test case. -Added some comments.

+++ src/main/java/libs/AverageTemperature.java
@@ -5,8 +5,15 @@ package libs;
  */
 public class AverageTemperature {
 
+  /*
+   * The original implementation will not work when input is huge as tsum
+   * uses byte (-128 to 128). When tsum exceeds max byte value, it would wrap
+   * around from min value. Use the data type with the biggest size as
+   * much as possible.
+   */
+
   public static double getTemperature(byte[] temperatures) {
-    int tsum = 0;
+    long tsum = 0;
     int i;
     for (i = 0; i < temperatures.length; i++) {
       tsum += temperatures[i];
